pip-requirements.txt
simple-signal-handler (0.1) 
re2 0.2.22
imaplib2 (2.45.0) 
argparse_config (0.5.1)
argparse_actions (0.4.4) 
pytz (2017.3) 
python-timeuuid (0.3.5) 
parsyfiles (2.4.2) 
uuid1 (1.0.1)
xsmtplib (1.0.0.dev1) 
secure-smtplib==0.1.1 
smtp-health-check==0.1.0
check_email_status==1.1.3  
dateutil.parser
miongo==0.1.1
rf-monitor==1.0.2
configparser==3.5.0 
pytz==2017.2
requests==2.18.4
